Merge from master (#8)
* Added Absorption effect

With some poorly-written lore and a slightly disgusting icon, but at
least the effect works.

* Discovered occulus descriptions are a thing now.

They're awesome. And I still can't write good lore.

* Hopefully fix block breaking in claims

i highly doubt this fixes it but hey, worth a shot

* Fix of broom texture

Changed the path from "textures/mobs/Broom.png" to "textures/mobs/broom.png" to fix texture.

* Fix of #47

* add ja_JP.lang

add Japanese Translation. (but incomplete

* 1.10.2:Fix render crash with conduits

* Prevent crash with essence refiner

Lots of issues with tile entities not being cleared wtf

* Fix fortune on dig spell

the start of fixing Prosperity.

* Fix ConcurrentModification exception with item drops

as items were being dropped, the loop was checking at the same time, causing CMEs to happen, simply copying the contents to a new array and use that temp array to iterate over everything fixes this.

* Update Bugs list

* Fix affinity not reducing mana cost

Fixes #60

* Fix broom despawning

Not sure if this will actually fix it but hey, worth a shot

* Possibly fix dryads not spawning

* Update bugs list